+++ rs/client-gen-v2/Cargo.toml
@@ -14,7 +14,7 @@ anyhow.workspace = true
 convert_case.workspace = true
 genco.workspace = true
 parity-scale-codec.workspace = true
-sails-idl-meta = { workspace = true, features = ["ast"] }
+sails-idl-ast.workspace = true
 sails-idl-parser-v2 = { workspace = true, features = ["std"] }
 
 [dev-dependencies]
