+++ rs/idl-ast/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "sails-idl-ast"
+description = "IDL AST model for the Sails framework"
+documentation = "https://docs.rs/sails-idl-ast"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+askama = { workspace = true, optional = true, features = ["alloc", "derive"] }
+keccak-const.workspace = true
+parity-scale-codec.workspace = true
+serde = { workspace = true, features = ["alloc", "derive"], optional = true }
+serde-json = { workspace = true, features = ["alloc"], optional = true }
+
+[dev-dependencies]
+gprimitives.workspace = true
+insta.workspace = true
+sails-reflect-hash.workspace = true
+
+[features]
+default = []
+serde = ["dep:serde", "dep:serde-json"]
+templates = ["dep:askama"]
+
+[[test]]
+name = "serde"
+required-features = ["serde"]
+
+[[test]]
+name = "templates"
+required-features = ["templates"]
