+++ rs/idl-ast/src/hash.rs
@@ -1,5 +1,5 @@
 use super::*;
-use alloc::collections::btree_map::BTreeMap;
+use alloc::{collections::btree_map::BTreeMap, format};
 use keccak_const::Keccak256;
 
 type Error = String;
