+++ rs/idl-ast/templates/struct_def.askama
