+++ rs/idl-ast/templates/variant.askama
