+++ rs/idl-ast/tests/templates.rs
@@ -1,5 +1,5 @@
 use askama::Template;
-use sails_idl_meta::*;
+use sails_idl_ast::*;
 
 mod fixture;
