+++ rs/idl-parser-v2/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "sails-idl-parser-v2"
 description = "IDL v2 parser for the Sails framework"
-documentation = "https://docs.rs/sails-idl-parser"
+documentation = "https://docs.rs/sails-idl-parser-v2"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -21,6 +21,6 @@ std = ["pest/std", "pest_derive/std", "thiserror/std", "dep:tempfile"]
 thiserror.workspace = true
 pest.workspace = true
 pest_derive.workspace = true
-sails-idl-meta = { workspace = true, features = ["ast"] }
+sails-idl-ast.workspace = true
 paste.workspace = true
 tempfile = { workspace = true, optional = true }
